% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_t1_data.R
\name{delete_t1_data}
\alias{delete_t1_data}
\title{Delete T1 Image Filenames}
\usage{
delete_t1_data(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{delete_kirby21_data}},
\code{modalities = "T1"} so it cannot be specified}
}
\value{
Nothing is returned
}
\description{
Delete the files for the T1 images
}
