% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_t1_data.R, R/download_t1_filenames.R
\name{download_t1_data}
\alias{download_t1_data}
\alias{download_t1_data}
\title{Download T1 Image Filenames}
\usage{
download_t1_data(...)

download_t1_data(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{download_kirby21_data}},
\code{modalities = "T1"} so it cannot be specified}

\item{...}{arguments to pass to \code{\link{download_kirby21_data}},
\code{modalities = "T1"} so it cannot be specified}
}
\value{
Indicator if the data is downloaded.

Indicator if the data is downloaded.
}
\description{
Download the files for the T1 images

Download the files for the T1 images
}
\examples{
download_t1_data()
download_t1_data()
}
